mega refactor merging leq and inversio -> first factory test passing 0_0
