+++ autoarray/dataset/interferometer.py
@@ -91,7 +91,7 @@ class SettingsInterferometer(AbstractSettingsDataset):
             The type of grid used to create the image from the `Galaxy` and `Plane`. The options are `Grid2D`,
             and `Grid2DIterate` (see the `Grid2D` documentation for a description of these options).
         grid_pixelized_class : ag.Grid2D
-            The type of grid used to create the grid that maps the `LEq` source pixels to the data's image-pixels.
+            The type of grid used to create the grid that maps the `Inversion` source pixels to the data's image-pixels.
             The options are `Grid2D` and `Grid2DIterate` (see the `Grid2D` documentation for a
             description of these options).
         sub_size
