+++ autoarray/inversion/inversion/imaging/inversion_imaging_util.py
@@ -1,8 +1,11 @@
 import numpy as np
+from scipy.linalg import cho_solve
 from typing import List, Optional, Tuple
 
-from autoarray import numba_util
+from autoarray.inversion.inversion.settings import SettingsInversion
 
+from autoarray import numba_util
+from autoarray import exc
 
 @numba_util.jit()
 def w_tilde_data_imaging_from(
@@ -435,80 +438,6 @@ def data_vector_via_blurred_mapping_matrix_from(
 
 
 @numba_util.jit()
-def data_vector_via_transformed_mapping_matrix_from(
-    transformed_mapping_matrix: np.ndarray,
-    visibilities: np.ndarray,
-    noise_map: np.ndarray,
-) -> np.ndarray:
-    """
-    Returns the data vector `D` from a transformed mapping matrix `f` and the 1D image `d` and 1D noise-map `sigma`
-    (see Warren & Dye 2003).
-
-    Parameters
-    -----------
-    transformed_mapping_matrix
-        The matrix representing the transformed mappings between sub-grid pixels and pixelization pixels.
-    image
-        Flattened 1D array of the observed image the inversion is fitting.
-    noise_map
-        Flattened 1D array of the noise-map used by the inversion during the fit.
-    """
-
-    data_vector = np.zeros(transformed_mapping_matrix.shape[1])
-
-    visibilities_real = visibilities.real
-    visibilities_imag = visibilities.imag
-    transformed_mapping_matrix_real = transformed_mapping_matrix.real
-    transformed_mapping_matrix_imag = transformed_mapping_matrix.imag
-    noise_map_real = noise_map.real
-    noise_map_imag = noise_map.imag
-
-    for vis_1d_index in range(transformed_mapping_matrix.shape[0]):
-        for pix_1d_index in range(transformed_mapping_matrix.shape[1]):
-            real_value = (
-                visibilities_real[vis_1d_index]
-                * transformed_mapping_matrix_real[vis_1d_index, pix_1d_index]
-                / (noise_map_real[vis_1d_index] ** 2.0)
-            )
-            imag_value = (
-                visibilities_imag[vis_1d_index]
-                * transformed_mapping_matrix_imag[vis_1d_index, pix_1d_index]
-                / (noise_map_imag[vis_1d_index] ** 2.0)
-            )
-            data_vector[pix_1d_index] += real_value + imag_value
-
-    return data_vector
-
-
-def curvature_matrix_via_w_tilde_from(
-    w_tilde: np.ndarray, mapping_matrix: np.ndarray
-) -> np.ndarray:
-    """
-    Returns the curvature matrix `F` (see Warren & Dye 2003) from `w_tilde`.
-
-    The dimensions of `w_tilde` are [image_pixels, image_pixels], meaning that for datasets with many image pixels
-    this matrix can take up 10's of GB of memory. The calculation of the `curvature_matrix` via this function will
-    therefore be very slow, and the method `curvature_matrix_via_w_tilde_curvature_preload_imaging_from` should be used
-    instead.
-
-    Parameters
-    ----------
-    w_tilde
-        A matrix of dimensions [image_pixels, image_pixels] that encodes the convolution or NUFFT of every image pixel
-        pair on the noise map.
-    mapping_matrix
-        The matrix representing the mappings between sub-grid pixels and pixelization pixels.
-
-    Returns
-    -------
-    ndarray
-        The curvature matrix `F` (see Warren & Dye 2003).
-    """
-
-    return np.dot(mapping_matrix.T, np.dot(w_tilde, mapping_matrix))
-
-
-@numba_util.jit()
 def curvature_matrix_via_w_tilde_curvature_preload_imaging_from(
     curvature_preload: np.ndarray,
     curvature_indexes: np.ndarray,
@@ -696,228 +625,3 @@ def curvature_matrix_off_diags_via_w_tilde_curvature_preload_imaging_from(
             curvature_index += 1
 
     return curvature_matrix
-
-
-def curvature_matrix_with_added_to_diag_from(
-    curvature_matrix: np.ndarray, linear_obj_func_index_list: Optional[List] = None
-) -> np.ndarray:
-    """
-    It is common for the `curvature_matrix` computed to not be positive-definite, leading for the inversion
-    via `np.linalg.solve` to fail and raise a `LinAlgError`.
-
-    In many circumstances, adding a small numerical value of `1.0e-8` to the diagonal of the `curvature_matrix`
-    makes it positive definite, such that the inversion is performed without raising an error.
-
-    This function adds this numerical value to the diagonal of the curvature matrix.
-
-    Parameters
-    ----------
-    curvature_matrix
-        The curvature matrix which is being constructed in order to solve a linear system of equations.
-    """
-
-    for i in linear_obj_func_index_list:
-        curvature_matrix[i, i] += 1e-8
-
-    return curvature_matrix
-
-
-def curvature_matrix_via_mapping_matrix_from(
-    mapping_matrix: np.ndarray,
-    noise_map: np.ndarray,
-    add_to_curvature_diag: bool = False,
-    linear_obj_func_index_list: Optional[List] = None,
-) -> np.ndarray:
-    """
-    Returns the curvature matrix `F` from a blurred mapping matrix `f` and the 1D noise-map $\sigma$
-     (see Warren & Dye 2003).
-
-    Parameters
-    -----------
-    mapping_matrix
-        The matrix representing the mappings (these could be blurred or transfomed) between sub-grid pixels and
-        pixelization pixels.
-    noise_map
-        Flattened 1D array of the noise-map used by the inversion during the fit.
-    """
-    array = mapping_matrix / noise_map[:, None]
-    curvature_matrix = np.dot(array.T, array)
-
-    if add_to_curvature_diag:
-        curvature_matrix = curvature_matrix_with_added_to_diag_from(
-            curvature_matrix=curvature_matrix,
-            linear_obj_func_index_list=linear_obj_func_index_list,
-        )
-
-    return curvature_matrix
-
-
-@numba_util.jit()
-def curvature_matrix_preload_from(
-    mapping_matrix: np.ndarray, mapping_matrix_threshold=1.0e-8
-) -> np.ndarray:
-    """
-    Returns a matrix that expresses the non-zero entries of the blurred mapping matrix for an efficient construction of
-    the curvature matrix `F` (see Warren & Dye 2003).
-
-    This is used for models where the blurred mapping matrix does not change, but the noise-map of the values that
-    are used to construct that blurred mapping matrix do change.
-
-    Parameters
-    -----------
-    mapping_matrix
-        The matrix representing the mappings (these could be blurred or transfomed) between sub-grid pixels and
-        pixelization pixels.
-    """
-
-    curvature_matrix_counts = np.zeros(mapping_matrix.shape[0])
-
-    for mask_1d_index in range(mapping_matrix.shape[0]):
-        for pix_index in range(mapping_matrix.shape[1]):
-            if mapping_matrix[mask_1d_index, pix_index] > mapping_matrix_threshold:
-                curvature_matrix_counts[mask_1d_index] += 1
-
-    preload_max = np.max(curvature_matrix_counts)
-
-    curvature_matrix_preload = np.zeros((mapping_matrix.shape[0], int(preload_max)))
-
-    for mask_1d_index in range(mapping_matrix.shape[0]):
-        index = 0
-        for pix_index in range(mapping_matrix.shape[1]):
-            if mapping_matrix[mask_1d_index, pix_index] > mapping_matrix_threshold:
-                curvature_matrix_preload[mask_1d_index, index] = pix_index
-                index += 1
-
-    return curvature_matrix_preload, curvature_matrix_counts
-
-
-@numba_util.jit()
-def curvature_matrix_via_sparse_preload_from(
-    mapping_matrix: np.ndarray,
-    noise_map: np.ndarray,
-    curvature_matrix_preload,
-    curvature_matrix_counts,
-) -> np.ndarray:
-    """
-    Returns the curvature matrix `F` from a blurred mapping matrix `f` and the 1D noise-map $\sigma$
-     (see Warren & Dye 2003) via a sparse preload matrix, which has already mapped out where all non-zero entries
-     and multiplications take place.
-
-    This is used for models where the blurred mapping matrix does not change, but the noise-map of the values that
-    are used to construct that blurred mapping matrix do change.
-
-    Parameters
-    -----------
-    mapping_matrix
-        The matrix representing the mappings (these could be blurred or transfomed) between sub-grid pixels and
-        pixelization pixels.
-    noise_map
-        Flattened 1D array of the noise-map used by the inversion during the fit.
-    """
-    curvature_matrix = np.zeros((mapping_matrix.shape[1], mapping_matrix.shape[1]))
-
-    for mask_1d_index in range(mapping_matrix.shape[0]):
-
-        total_pix = curvature_matrix_counts[mask_1d_index]
-
-        for preload_index_0 in range(total_pix):
-            for preload_index_1 in range(
-                preload_index_0, curvature_matrix_counts[mask_1d_index]
-            ):
-
-                pix_index_0 = curvature_matrix_preload[mask_1d_index, preload_index_0]
-                pix_index_1 = curvature_matrix_preload[mask_1d_index, preload_index_1]
-
-                curvature_matrix[pix_index_0, pix_index_1] += (
-                    mapping_matrix[mask_1d_index, pix_index_0]
-                    * mapping_matrix[mask_1d_index, pix_index_1]
-                ) / noise_map[mask_1d_index] ** 2
-
-    for i in range(mapping_matrix.shape[1]):
-        for j in range(mapping_matrix.shape[1]):
-            curvature_matrix[j, i] = curvature_matrix[i, j]
-
-    return curvature_matrix
-
-
-@numba_util.jit()
-def mapped_reconstructed_data_via_image_to_pix_unique_from(
-    data_to_pix_unique: np.ndarray,
-    data_weights: np.ndarray,
-    pix_lengths: np.ndarray,
-    reconstruction: np.ndarray,
-) -> np.ndarray:
-    """
-    Returns the reconstructed data vector from the blurred mapping matrix `f` and solution vector *S*.
-
-    Parameters
-    -----------
-    mapping_matrix
-        The matrix representing the blurred mappings between sub-grid pixels and pixelization pixels.
-
-    """
-
-    data_pixels = data_to_pix_unique.shape[0]
-
-    mapped_reconstructed_data = np.zeros(data_pixels)
-
-    for data_0 in range(data_pixels):
-        for pix_0 in range(pix_lengths[data_0]):
-
-            pix_for_data = data_to_pix_unique[data_0, pix_0]
-
-            mapped_reconstructed_data[data_0] += (
-                data_weights[data_0, pix_0] * reconstruction[pix_for_data]
-            )
-
-    return mapped_reconstructed_data
-
-
-@numba_util.jit()
-def mapped_reconstructed_data_via_mapping_matrix_from(
-    mapping_matrix: np.ndarray, reconstruction: np.ndarray
-) -> np.ndarray:
-    """
-    Returns the reconstructed data vector from the blurred mapping matrix `f` and solution vector *S*.
-
-    Parameters
-    -----------
-    mapping_matrix
-        The matrix representing the blurred mappings between sub-grid pixels and pixelization pixels.
-
-    """
-    mapped_reconstructed_data = np.zeros(mapping_matrix.shape[0])
-    for i in range(mapping_matrix.shape[0]):
-        for j in range(reconstruction.shape[0]):
-            mapped_reconstructed_data[i] += reconstruction[j] * mapping_matrix[i, j]
-
-    return mapped_reconstructed_data
-
-
-@numba_util.jit()
-def mapped_reconstructed_visibilities_from(
-    transformed_mapping_matrix: np.ndarray, reconstruction: np.ndarray
-) -> np.ndarray:
-    """
-    Returns the reconstructed data vector from the blurrred mapping matrix `f` and solution vector *S*.
-
-    Parameters
-    -----------
-    transformed_mapping_matrix
-        The matrix representing the blurred mappings between sub-grid pixels and pixelization pixels.
-
-    """
-    mapped_reconstructed_visibilities = (0.0 + 0.0j) * np.zeros(
-        transformed_mapping_matrix.shape[0]
-    )
-
-    transformed_mapping_matrix_real = transformed_mapping_matrix.real
-    transformed_mapping_matrix_imag = transformed_mapping_matrix.imag
-
-    for i in range(transformed_mapping_matrix.shape[0]):
-        for j in range(reconstruction.shape[0]):
-            mapped_reconstructed_visibilities[i] += (
-                reconstruction[j] * transformed_mapping_matrix_real[i, j]
-            ) + 1.0j * (reconstruction[j] * transformed_mapping_matrix_imag[i, j])
-
-    return mapped_reconstructed_visibilities
