+++ autoarray/inversion/inversion/interferometer/w_tilde.py
@@ -3,42 +3,46 @@ from typing import Dict, List, Optional
 
 from autoconf import cached_property
 
-from autoarray.inversion.linear_eqn.interferometer.abstract import (
-    AbstractLEqInterferometer,
+from autoarray.inversion.inversion.interferometer.abstract import (
+    AbstractInversionInterferometer,
 )
+from autoarray.dataset.interferometer import WTildeInterferometer
 from autoarray.inversion.linear_obj.func_list import LinearObj
+from autoarray.inversion.linear_obj.linear_obj_reg import LinearObjReg
 from autoarray.inversion.inversion.settings import SettingsInversion
 from autoarray.preloads import Preloads
 from autoarray.operators.transformer import TransformerNUFFT
-from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.visibilities import Visibilities
 from autoarray.structures.visibilities import VisibilitiesNoiseMap
 
-from autoarray.inversion.linear_eqn import leq_util
+from autoarray.inversion.inversion import inversion_util
 
 from autoarray.numba_util import profile_func
 
 
-class LEqInterferometerMapping(AbstractLEqInterferometer):
+class InversionInterferometerWTilde(AbstractInversionInterferometer):
     def __init__(
         self,
+        data: Visibilities,
         noise_map: VisibilitiesNoiseMap,
         transformer: TransformerNUFFT,
-        linear_obj_list: List[LinearObj],
+        w_tilde: WTildeInterferometer,
+        linear_obj_reg_list: List[LinearObjReg],
         settings: SettingsInversion = SettingsInversion(),
+        preloads: Preloads = Preloads(),
         profiling_dict: Optional[Dict] = None,
     ):
         """
         Constructs linear equations (via vectors and matrices) which allow for sets of simultaneous linear equations
-        to be solved (see `inversion.linear_eqn.abstract.AbstractLEq` for a full description).
+        to be solved (see `inversion.inversion.abstract.AbstractInversion` for a full description).
 
         A linear object describes the mappings between values in observed `data` and the linear object's model via its
         `mapping_matrix`. This class constructs linear equations for `Interferometer` objects, where the data is an
         an array of visibilities and the mappings include a non-uniform fast Fourier transform operation described by
         the interferometer dataset's transformer.
 
-        This class uses the mapping formalism, which constructs the simultaneous linear equations using the
-        `mapping_matrix` of every linear object.
+        This class uses the w-tilde formalism, which speeds up the construction of the simultaneous linear equations by
+        bypassing the construction of a `mapping_matrix`.
 
         Parameters
         -----------
@@ -47,23 +51,34 @@ class LEqInterferometerMapping(AbstractLEqInterferometer):
         transformer
             The transformer which performs a non-uniform fast Fourier transform operations on the mapping matrix
             with the interferometer data's transformer.
-        linear_obj_list
+        w_tilde
+            An object containing matrices that construct the linear equations via the w-tilde formalism which bypasses
+            the mapping matrix.
+        linear_obj_reg_list
             The linear objects used to reconstruct the data's observed values. If multiple linear objects are passed
             the simultaneous linear equations are combined and solved simultaneously.
         profiling_dict
             A dictionary which contains timing of certain functions calls which is used for profiling.
         """
 
+        self.w_tilde = w_tilde
+        self.w_tilde.check_noise_map(noise_map=noise_map)
+
         super().__init__(
+            data=data,
             noise_map=noise_map,
             transformer=transformer,
-            linear_obj_list=linear_obj_list,
+            linear_obj_reg_list=linear_obj_reg_list,
             settings=settings,
+            preloads=preloads,
             profiling_dict=profiling_dict,
         )
 
+        self.settings = settings
+
+    @cached_property
     @profile_func
-    def data_vector_from(self, data: Array2D, preloads: Preloads) -> np.ndarray:
+    def data_vector(self) -> np.ndarray:
         """
         The `data_vector` is a 1D vector whose values are solved for by the simultaneous linear equations constructed
         by this object.
@@ -71,15 +86,13 @@ class LEqInterferometerMapping(AbstractLEqInterferometer):
         The linear algebra is described in the paper https://arxiv.org/pdf/astro-ph/0302587.pdf), where the
         data vector is given by equation (4) and the letter D.
 
-        If there are multiple linear objects their `operated_mapping_matrix` properties will have already been
-        concatenated ensuring their `data_vector` values are solved for simultaneously.
+        If there are multiple linear objects the `data_vectors` are concatenated ensuring their values are solved
+        for simultaneously.
 
-        The calculation is described in more detail in `leq_util.data_vector_via_transformed_mapping_matrix_from`.
+        The calculation is described in more detail in `inversion_util.w_tilde_data_interferometer_from`.
         """
-        return leq_util.data_vector_via_transformed_mapping_matrix_from(
-            transformed_mapping_matrix=self.operated_mapping_matrix,
-            visibilities=data,
-            noise_map=self.noise_map,
+        return np.dot(
+            self.linear_obj_list[0].mapping_matrix.T, self.w_tilde.dirty_image
         )
 
     @cached_property
@@ -96,30 +109,48 @@ class LEqInterferometerMapping(AbstractLEqInterferometer):
         concatenated ensuring their `curvature_matrix` values are solved for simultaneously. This includes all
         diagonal and off-diagonal terms describing the covariances between linear objects.
         """
+        return self.curvature_matrix_diag
 
-        real_curvature_matrix = leq_util.curvature_matrix_via_mapping_matrix_from(
-            mapping_matrix=self.operated_mapping_matrix.real,
-            noise_map=self.noise_map.real,
-        )
+    @property
+    @profile_func
+    def curvature_matrix_diag(self) -> np.ndarray:
+        """
+        The `curvature_matrix` is a 2D matrix which uses the mappings between the data and the linear objects to
+        construct the simultaneous linear equations.
 
-        imag_curvature_matrix = leq_util.curvature_matrix_via_mapping_matrix_from(
-            mapping_matrix=self.operated_mapping_matrix.imag,
-            noise_map=self.noise_map.imag,
-        )
+        The linear algebra is described in the paper https://arxiv.org/pdf/astro-ph/0302587.pdf, where the
+        curvature matrix given by equation (4) and the letter F.
 
-        curvature_matrix = np.add(real_curvature_matrix, imag_curvature_matrix)
+        This function computes the diagonal terms of F using the w_tilde formalism.
+        """
 
-        if self.add_to_curvature_diag:
-            curvature_matrix = leq_util.curvature_matrix_with_added_to_diag_from(
-                curvature_matrix=curvature_matrix,
-                linear_obj_func_index_list=self.linear_obj_func_index_list,
+        if self.settings.use_w_tilde_numpy:
+
+            return inversion_util.curvature_matrix_via_w_tilde_from(
+                w_tilde=self.w_tilde.w_matrix, mapping_matrix=self.mapping_matrix
             )
 
-        return curvature_matrix
+        from autoarray.inversion.inversion import inversion_util_secret
 
+        return inversion_util_secret.curvature_matrix_via_w_tilde_curvature_preload_interferometer_from(
+            curvature_preload=self.w_tilde.curvature_preload,
+            pix_indexes_for_sub_slim_index=self.mapper_list[
+                0
+            ].pix_indexes_for_sub_slim_index,
+            pix_size_for_sub_slim_index=self.mapper_list[
+                0
+            ].pix_sizes_for_sub_slim_index,
+            pix_weights_for_sub_slim_index=self.mapper_list[
+                0
+            ].pix_weights_for_sub_slim_index,
+            native_index_for_slim_index=self.transformer.real_space_mask.native_index_for_slim_index,
+            pix_pixels=self.linear_obj_reg_list[0].pixels,
+        )
+
+    @property
     @profile_func
-    def mapped_reconstructed_data_dict_from(
-        self, reconstruction: np.ndarray
+    def mapped_reconstructed_data_dict(
+        self,
     ) -> Dict[LinearObj, Visibilities]:
         """
         When constructing the simultaneous linear equations (via vectors and matrices) the quantities of each individual
@@ -146,19 +177,12 @@ class LEqInterferometerMapping(AbstractLEqInterferometer):
         """
         mapped_reconstructed_data_dict = {}
 
-        reconstruction_dict = self.source_quantity_dict_from(
-            source_quantity=reconstruction
-        )
+        image_dict = self.mapped_reconstructed_image_dict
 
-        operated_mapping_matrix_list = self.operated_mapping_matrix_list
+        for linear_obj in self.linear_obj_list:
 
-        for index, linear_obj in enumerate(self.linear_obj_list):
-
-            reconstruction = reconstruction_dict[linear_obj]
-
-            visibilities = leq_util.mapped_reconstructed_visibilities_from(
-                transformed_mapping_matrix=operated_mapping_matrix_list[index],
-                reconstruction=reconstruction,
+            visibilities = self.transformer.visibilities_from(
+                image=image_dict[linear_obj]
             )
 
             visibilities = Visibilities(visibilities=visibilities)
