+++ autoarray/inversion/mappers/abstract.py
@@ -197,7 +197,7 @@ class AbstractMapper(LinearObj):
 
     @cached_property
     @profile_func
-    def data_unique_mappings(self) -> "UniqueMappings":
+    def unique_mappings(self) -> "UniqueMappings":
         """
         Returns the unique mappings of every unmasked data pixel's (e.g. `grid_slim`) sub-pixels (e.g. `grid_sub_slim`)
         to their corresponding pixelization pixels (e.g. `pixelization_grid`).
