+++ autoarray/util/__init__.py
@@ -12,6 +12,6 @@ from autoarray.fit import fit_util as fit
 from autoarray.inversion.pixelizations import pixelization_util as pixelization
 from autoarray.inversion.mappers import mapper_util as mapper
 from autoarray.inversion.regularization import regularization_util as regularization
-from autoarray.inversion.linear_eqn import leq_util as leq
+from autoarray.inversion.inversion import inversion_util as inversion
 from autoarray.inversion.inversion import inversion_util as inversion
 from autoarray.operators import transformer_util as transformer
