+++ test_autoarray/fit/test_fit_dataset.py
@@ -5,10 +5,10 @@ import autoarray as aa
 
 def test__inversion_figure_of_merit(masked_imaging_7x7, model_image_7x7):
 
-    leq = aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()], data_vector=1)
+    inversion = aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()], data_vector=1)
 
     inversion = aa.m.MockInversion(
-        leq=leq,
+        inversion=inversion,
         regularization_term=2.0,
         log_det_curvature_reg_matrix_term=3.0,
         log_det_regularization_matrix_term=4.0,
@@ -23,10 +23,10 @@ def test__inversion_figure_of_merit(masked_imaging_7x7, model_image_7x7):
 
     assert fit.figure_of_merit == fit.log_evidence
 
-    leq = aa.m.MockLEq(linear_obj_list=[aa.m.MockLinearObjFunc()], data_vector=1)
+    inversion = aa.m.MockInversion(linear_obj_list=[aa.m.MockLinearObjFunc()], data_vector=1)
 
     inversion = aa.m.MockInversion(
-        leq=leq,
+        inversion=inversion,
         regularization_term=2.0,
         log_det_curvature_reg_matrix_term=3.0,
         log_det_regularization_matrix_term=4.0,
