+++ test_autoarray/inversion/inversion/test_abstract.py
@@ -63,7 +63,7 @@ def test__has_regularization():
 def test__linear_obj_func_list__filters_other_objects():
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(
+        inversion=aa.m.MockInversion(
             linear_obj_list=[
                 aa.m.MockMapper(pixels=1),
                 aa.m.MockLinearObjFunc(grid=3),
@@ -81,7 +81,7 @@ def test__linear_obj_func_list__filters_other_objects():
 def test__mapper_list__filters_other_objects():
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(
+        inversion=aa.m.MockInversion(
             linear_obj_list=[
                 aa.m.MockMapper(pixels=1),
                 aa.m.MockLinearObjFunc(),
@@ -97,7 +97,7 @@ def test__mapper_list__filters_other_objects():
 
 def test__regularization_matrix():
 
-    leq = aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper(), aa.m.MockMapper()])
+    inversion = aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper(), aa.m.MockMapper()])
 
     reg_0 = aa.m.MockRegularization(regularization_matrix=np.ones((2, 2)))
     reg_1 = aa.m.MockRegularization(regularization_matrix=2.0 * np.ones((3, 3)))
@@ -106,7 +106,7 @@ def test__regularization_matrix():
     linear_obj_reg_1 = aa.m.MockLinearObjReg(regularization=reg_1)
     #
     inversion = aa.m.MockInversion(
-        linear_obj_reg_list=[linear_obj_reg_0, linear_obj_reg_1], leq=leq
+        linear_obj_reg_list=[linear_obj_reg_0, linear_obj_reg_1], inversion=inversion
     )
 
     regularization_matrix = np.array(
@@ -126,7 +126,7 @@ def test__preloads__operated_mapping_matrix_and_curvature_matrix_preload():
 
     operated_mapping_matrix = 2.0 * np.ones((9, 3))
 
-    curvature_matrix_preload, curvature_matrix_counts = aa.util.leq.curvature_matrix_preload_from(
+    curvature_matrix_preload, curvature_matrix_counts = aa.util.inversion.curvature_matrix_preload_from(
         mapping_matrix=operated_mapping_matrix
     )
 
@@ -137,9 +137,9 @@ def test__preloads__operated_mapping_matrix_and_curvature_matrix_preload():
     )
 
     # noinspection PyTypeChecker
-    leq = aa.m.MockLEq(noise_map=np.ones(9), linear_obj_list=aa.m.MockMapper())
+    inversion = aa.m.MockInversion(noise_map=np.ones(9), linear_obj_list=aa.m.MockMapper())
 
-    inversion = aa.m.MockInversion(leq=leq, preloads=preloads)
+    inversion = aa.m.MockInversion(inversion=inversion, preloads=preloads)
 
     assert inversion.operated_mapping_matrix[0, 0] == 2.0
     assert inversion.curvature_matrix[0, 0] == 36.0
@@ -162,7 +162,7 @@ def test__reconstruction_dict():
     mapper = aa.m.MockMapper(pixels=3)
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(linear_obj_list=[linear_obj, mapper]),
+        inversion=aa.m.MockInversion(linear_obj_list=[linear_obj, mapper]),
         reconstruction=reconstruction,
     )
 
@@ -175,7 +175,7 @@ def test__reconstruction_dict():
     mapper_1 = aa.m.MockMapper(pixels=3)
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(linear_obj_list=[linear_obj, mapper_0, mapper_1]),
+        inversion=aa.m.MockInversion(linear_obj_list=[linear_obj, mapper_0, mapper_1]),
         reconstruction=reconstruction,
     )
 
@@ -191,7 +191,7 @@ def test__mapped_reconstructed_data():
 
     # noinspection PyTypeChecker
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(mapped_reconstructed_data_dict=mapped_reconstructed_data_dict),
+        inversion=aa.m.MockInversion(mapped_reconstructed_data_dict=mapped_reconstructed_data_dict),
         reconstruction=np.ones(3),
         reconstruction_dict=[None],
     )
@@ -208,7 +208,7 @@ def test__mapped_reconstructed_data():
 
     # noinspection PyTypeChecker
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(mapped_reconstructed_data_dict=mapped_reconstructed_data_dict),
+        inversion=aa.m.MockInversion(mapped_reconstructed_data_dict=mapped_reconstructed_data_dict),
         reconstruction=np.array([1.0, 1.0, 2.0, 2.0]),
         reconstruction_dict=[None, None],
     )
@@ -228,7 +228,7 @@ def test__mapped_reconstructed_image():
 
     # noinspection PyTypeChecker
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(
+        inversion=aa.m.MockInversion(
             mapped_reconstructed_image_dict=mapped_reconstructed_image_dict
         ),
         reconstruction=np.ones(3),
@@ -247,7 +247,7 @@ def test__mapped_reconstructed_image():
 
     # noinspection PyTypeChecker
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(
+        inversion=aa.m.MockInversion(
             mapped_reconstructed_image_dict=mapped_reconstructed_image_dict
         ),
         reconstruction=np.array([1.0, 1.0, 2.0, 2.0]),
@@ -282,10 +282,10 @@ def test__regularization_term():
         [[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]]
     )
 
-    leq = aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()])
+    inversion = aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()])
 
     inversion = aa.m.MockInversion(
-        leq=leq,
+        inversion=inversion,
         reconstruction=reconstruction,
         linear_obj_reg_list=[aa.m.MockLinearObjReg(regularization=1, pixels=3)],
         regularization_matrix=regularization_matrix,
@@ -313,10 +313,10 @@ def test__regularization_term():
         [[2.0, -1.0, 0.0], [-1.0, 2.0, -1.0], [0.0, -1.0, 2.0]]
     )
 
-    leq = aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()])
+    inversion = aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()])
 
     inversion = aa.m.MockInversion(
-        leq=leq,
+        inversion=inversion,
         reconstruction=reconstruction,
         linear_obj_reg_list=[aa.m.MockLinearObjReg(regularization=1, pixels=3)],
         regularization_matrix=regularization_matrix,
@@ -341,10 +341,10 @@ def test__regularization_term():
 
 def test__preload_of_log_det_regularization_term_overwrites_calculation():
 
-    leq = aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()])
+    inversion = aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()])
 
     inversion = aa.m.MockInversion(
-        leq=leq,
+        inversion=inversion,
         linear_obj_reg_list=[aa.m.MockLinearObjReg(regularization=1, pixels=3)],
         preloads=aa.Preloads(log_det_regularization_matrix_term=1.0),
     )
@@ -356,21 +356,21 @@ def test__determinant_of_positive_definite_matrix_via_cholesky():
 
     matrix = np.array([[1.0, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, 1.0]])
 
-    leq = aa.m.MockLEq(linear_obj_list=[None])
+    inversion = aa.m.MockInversion(linear_obj_list=[None])
 
-    leq = aa.m.MockInversion(leq=leq, curvature_reg_matrix=matrix)
+    inversion = aa.m.MockInversion(inversion=inversion, curvature_reg_matrix=matrix)
 
     log_determinant = np.log(np.linalg.det(matrix))
 
-    assert log_determinant == pytest.approx(leq.log_det_curvature_reg_matrix_term, 1e-4)
+    assert log_determinant == pytest.approx(inversion.log_det_curvature_reg_matrix_term, 1e-4)
 
     matrix = np.array([[2.0, -1.0, 0.0], [-1.0, 2.0, -1.0], [0.0, -1.0, 2.0]])
 
-    leq = aa.m.MockInversion(leq=leq, curvature_reg_matrix=matrix)
+    inversion = aa.m.MockInversion(inversion=inversion, curvature_reg_matrix=matrix)
 
     log_determinant = np.log(np.linalg.det(matrix))
 
-    assert log_determinant == pytest.approx(leq.log_det_curvature_reg_matrix_term, 1e-4)
+    assert log_determinant == pytest.approx(inversion.log_det_curvature_reg_matrix_term, 1e-4)
 
 
 def test__brightest_reconstruction_pixel_and_centre():
@@ -385,7 +385,7 @@ def test__brightest_reconstruction_pixel_and_centre():
     )
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(linear_obj_list=[mapper]),
+        inversion=aa.m.MockInversion(linear_obj_list=[mapper]),
         reconstruction=np.array([2.0, 3.0, 5.0, 0.0]),
     )
 
@@ -403,7 +403,7 @@ def test__interpolated_reconstruction_list_from():
     mapper = aa.m.MockMapper(pixels=3, interpolated_array=interpolated_array)
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(linear_obj_list=[mapper]), reconstruction=interpolated_array
+        inversion=aa.m.MockInversion(linear_obj_list=[mapper]), reconstruction=interpolated_array
     )
 
     interpolated_reconstruction_list = inversion.interpolated_reconstruction_list_from(
@@ -420,7 +420,7 @@ def test__interpolated_errors_list_from():
     mapper = aa.m.MockMapper(pixels=3, interpolated_array=interpolated_array)
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(linear_obj_list=[mapper]), errors=interpolated_array
+        inversion=aa.m.MockInversion(linear_obj_list=[mapper]), errors=interpolated_array
     )
 
     interpolated_errors_list = inversion.interpolated_errors_list_from(
