+++ test_autoarray/inversion/linear_eqn/test_imaging.py
@@ -1,6 +1,6 @@
 import autoarray as aa
 from autoarray.dataset.imaging import WTildeImaging
-from autoarray.inversion.linear_eqn.imaging.w_tilde import LEqImagingWTilde
+from autoarray.inversion.inversion.imaging.w_tilde import InversionImagingWTilde
 
 from autoarray import exc
 
@@ -13,16 +13,16 @@ directory = path.dirname(path.realpath(__file__))
 
 def test__operated_mapping_matrix_property(convolver_7x7, rectangular_mapper_7x7_3x3):
 
-    leq = aa.m.MockLEqImaging(
+    inversion = aa.m.MockInversionImaging(
         convolver=convolver_7x7, linear_obj_list=[rectangular_mapper_7x7_3x3]
     )
 
-    assert leq.operated_mapping_matrix_list[0][0, 0] == pytest.approx(1.0, 1e-4)
-    assert leq.operated_mapping_matrix[0, 0] == pytest.approx(1.0, 1e-4)
+    assert inversion.operated_mapping_matrix_list[0][0, 0] == pytest.approx(1.0, 1e-4)
+    assert inversion.operated_mapping_matrix[0, 0] == pytest.approx(1.0, 1e-4)
 
     convolver = aa.m.MockConvolver(operated_mapping_matrix=np.ones((2, 2)))
 
-    leq = aa.m.MockLEqImaging(
+    inversion = aa.m.MockInversionImaging(
         convolver=convolver,
         linear_obj_list=[rectangular_mapper_7x7_3x3, rectangular_mapper_7x7_3x3],
     )
@@ -31,13 +31,13 @@ def test__operated_mapping_matrix_property(convolver_7x7, rectangular_mapper_7x7
     operated_mapping_matrix_1 = np.array([[1.0, 1.0], [1.0, 1.0]])
     operated_mapping_matrix = np.array([[1.0, 1.0, 1.0, 1.0], [1.0, 1.0, 1.0, 1.0]])
 
-    assert leq.operated_mapping_matrix_list[0] == pytest.approx(
+    assert inversion.operated_mapping_matrix_list[0] == pytest.approx(
         operated_mapping_matrix_0, 1.0e-4
     )
-    assert leq.operated_mapping_matrix_list[1] == pytest.approx(
+    assert inversion.operated_mapping_matrix_list[1] == pytest.approx(
         operated_mapping_matrix_1, 1.0e-4
     )
-    assert leq.operated_mapping_matrix == pytest.approx(operated_mapping_matrix, 1.0e-4)
+    assert inversion.operated_mapping_matrix == pytest.approx(operated_mapping_matrix, 1.0e-4)
 
 
 def test__operated_mapping_matrix_property__with_operated_mapping_matrix_override(
@@ -53,20 +53,20 @@ def test__operated_mapping_matrix_property__with_operated_mapping_matrix_overrid
         operated_mapping_matrix_override=operated_mapping_matrix_override,
     )
 
-    leq = aa.m.MockLEqImaging(
+    inversion = aa.m.MockInversionImaging(
         convolver=convolver, linear_obj_list=[rectangular_mapper_7x7_3x3, linear_obj]
     )
 
     operated_mapping_matrix_0 = np.array([[1.0, 1.0], [1.0, 1.0]])
     operated_mapping_matrix = np.array([[1.0, 1.0, 1.0, 2.0], [1.0, 1.0, 3.0, 4.0]])
 
-    assert leq.operated_mapping_matrix_list[0] == pytest.approx(
+    assert inversion.operated_mapping_matrix_list[0] == pytest.approx(
         operated_mapping_matrix_0, 1.0e-4
     )
-    assert leq.operated_mapping_matrix_list[1] == pytest.approx(
+    assert inversion.operated_mapping_matrix_list[1] == pytest.approx(
         operated_mapping_matrix_override, 1.0e-4
     )
-    assert leq.operated_mapping_matrix == pytest.approx(operated_mapping_matrix, 1.0e-4)
+    assert inversion.operated_mapping_matrix == pytest.approx(operated_mapping_matrix, 1.0e-4)
 
 
 def test__curvature_matrix(rectangular_mapper_7x7_3x3):
@@ -81,29 +81,29 @@ def test__curvature_matrix(rectangular_mapper_7x7_3x3):
         operated_mapping_matrix_override=operated_mapping_matrix_override,
     )
 
-    leq = aa.LEqImagingMapping(
+    inversion = aa.InversionImagingMapping(
         linear_obj_list=[linear_obj, rectangular_mapper_7x7_3x3],
         noise_map=noise_map,
         convolver=convolver,
-        settings=aa.SettingsInversion(linear_funcs_add_to_curvature_diag=False),
+        settings=aa.SettingsInversion(no_regularization_add_to_curvature_diag=False),
     )
 
-    assert leq.curvature_matrix[0:2, 0:2] == pytest.approx(
+    assert inversion.curvature_matrix[0:2, 0:2] == pytest.approx(
         np.array([[10.0, 14.0], [14.0, 20.0]]), 1.0e-4
     )
 
-    assert leq.curvature_matrix[0, 0] - 10.0 < 1.0e-12
-    assert leq.curvature_matrix[3, 3] - 2.0 < 1.0e-12
+    assert inversion.curvature_matrix[0, 0] - 10.0 < 1.0e-12
+    assert inversion.curvature_matrix[3, 3] - 2.0 < 1.0e-12
 
-    leq = aa.LEqImagingMapping(
+    inversion = aa.InversionImagingMapping(
         linear_obj_list=[linear_obj, rectangular_mapper_7x7_3x3],
         noise_map=noise_map,
         convolver=convolver,
-        settings=aa.SettingsInversion(linear_funcs_add_to_curvature_diag=True),
+        settings=aa.SettingsInversion(no_regularization_add_to_curvature_diag=True),
     )
 
-    assert leq.curvature_matrix[0, 0] - 10.0 > 0.0
-    assert leq.curvature_matrix[3, 3] - 2.0 < 1.0e-12
+    assert inversion.curvature_matrix[0, 0] - 10.0 > 0.0
+    assert inversion.curvature_matrix[3, 3] - 2.0 < 1.0e-12
 
 
 def test__w_tilde_checks_noise_map_and_raises_exception_if_preloads_dont_match_noise_map():
@@ -125,7 +125,7 @@ def test__w_tilde_checks_noise_map_and_raises_exception_if_preloads_dont_match_n
     with pytest.raises(exc.InversionException):
 
         # noinspection PyTypeChecker
-        LEqImagingWTilde(
+        InversionImagingWTilde(
             noise_map=np.ones(9),
             convolver=aa.m.MockConvolver(matrix_shape),
             w_tilde=w_tilde,
