+++ test_autoarray/test_preloads.py
@@ -19,7 +19,7 @@ def test__set_w_tilde():
     # Noise maps of fit are different but there is an inversion, so we should not preload w_tilde and use w_tilde.
 
     inversion = aa.m.MockInversion(
-        leq=aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()])
+        inversion=aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()])
     )
 
     fit_0 = aa.m.MockFitImaging(
@@ -57,7 +57,7 @@ def test__set_w_tilde():
     preloads = aa.Preloads(w_tilde=1, use_w_tilde=1)
     preloads.set_w_tilde_imaging(fit_0=fit_0, fit_1=fit_1)
 
-    curvature_preload, indexes, lengths = aa.util.leq.w_tilde_curvature_preload_imaging_from(
+    curvature_preload, indexes, lengths = aa.util.inversion.w_tilde_curvature_preload_imaging_from(
         noise_map_native=fit_0.noise_map.native,
         kernel_native=fit_0.dataset.psf.native,
         native_index_for_slim_index=fit_0.dataset.mask.native_index_for_slim_index,
@@ -72,7 +72,7 @@ def test__set_w_tilde():
 
 def test__set_relocated_grid():
 
-    # LEq is None so there is no mapper, thus preload mapper to None.
+    # Inversion is None so there is no mapper, thus preload mapper to None.
 
     fit_0 = aa.m.MockFitImaging(inversion=None)
     fit_1 = aa.m.MockFitImaging(inversion=None)
@@ -84,15 +84,15 @@ def test__set_relocated_grid():
 
     # Mapper's mapping matrices are different, thus preload mapper to None.
 
-    leq_0 = aa.m.MockLEq(
+    inversion_0 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(source_grid_slim=np.ones((3, 2)))]
     )
-    leq_1 = aa.m.MockLEq(
+    inversion_1 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(source_grid_slim=2.0 * np.ones((3, 2)))]
     )
 
-    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_0))
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_0))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(relocated_grid=1)
     preloads.set_relocated_grid(fit_0=fit_0, fit_1=fit_1)
@@ -101,15 +101,15 @@ def test__set_relocated_grid():
 
     # Mapper's mapping matrices are the same, thus preload mapper.
 
-    leq_0 = aa.m.MockLEq(
+    inversion_0 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(source_grid_slim=np.ones((3, 2)))]
     )
-    leq_1 = aa.m.MockLEq(
+    inversion_1 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(source_grid_slim=np.ones((3, 2)))]
     )
 
-    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_0))
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_0))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(relocated_grid=1)
     preloads.set_relocated_grid(fit_0=fit_0, fit_1=fit_1)
@@ -119,7 +119,7 @@ def test__set_relocated_grid():
 
 def test__set_mapper_list():
 
-    # LEq is None so there is no mapper, thus preload mapper to None.
+    # Inversion is None so there is no mapper, thus preload mapper to None.
 
     fit_0 = aa.m.MockFitImaging(inversion=None)
     fit_1 = aa.m.MockFitImaging(inversion=None)
@@ -131,15 +131,15 @@ def test__set_mapper_list():
 
     # Mapper's mapping matrices are different, thus preload mapper to None.
 
-    leq_0 = aa.m.MockLEq(
+    inversion_0 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(mapping_matrix=np.ones((3, 2)))]
     )
-    leq_1 = aa.m.MockLEq(
+    inversion_1 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(mapping_matrix=2.0 * np.ones((3, 2)))]
     )
 
-    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_0))
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_0))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(mapper_list=1)
     preloads.set_mapper_list(fit_0=fit_0, fit_1=fit_1)
@@ -148,15 +148,15 @@ def test__set_mapper_list():
 
     # Mapper's mapping matrices are the same, thus preload mapper.
 
-    leq_0 = aa.m.MockLEq(
+    inversion_0 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(mapping_matrix=np.ones((3, 2)))]
     )
-    leq_1 = aa.m.MockLEq(
+    inversion_1 = aa.m.MockInversion(
         linear_obj_list=[aa.m.MockMapper(mapping_matrix=np.ones((3, 2)))]
     )
 
-    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_0))
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_0))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(mapper_list=1)
     preloads.set_mapper_list(fit_0=fit_0, fit_1=fit_1)
@@ -165,21 +165,21 @@ def test__set_mapper_list():
 
     # Multiple mappers pre inversion still preloads full mapper list.
 
-    leq_0 = aa.m.MockLEq(
+    inversion_0 = aa.m.MockInversion(
         linear_obj_list=[
             aa.m.MockMapper(mapping_matrix=np.ones((3, 2))),
             aa.m.MockMapper(mapping_matrix=np.ones((3, 2))),
         ]
     )
-    leq_1 = aa.m.MockLEq(
+    inversion_1 = aa.m.MockInversion(
         linear_obj_list=[
             aa.m.MockMapper(mapping_matrix=np.ones((3, 2))),
             aa.m.MockMapper(mapping_matrix=np.ones((3, 2))),
         ]
     )
 
-    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_0))
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_0))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(mapper_list=1)
     preloads.set_mapper_list(fit_0=fit_0, fit_1=fit_1)
@@ -193,7 +193,7 @@ def test__set_operated_mapping_matrix_with_preloads():
     curvature_matrix_preload = np.array([[1.0]])
     curvature_matrix_counts = np.array([1.0])
 
-    # LEq is None thus preload it to None.
+    # Inversion is None thus preload it to None.
 
     fit_0 = aa.m.MockFitImaging(inversion=None)
     fit_1 = aa.m.MockFitImaging(inversion=None)
@@ -209,7 +209,7 @@ def test__set_operated_mapping_matrix_with_preloads():
     assert preloads.curvature_matrix_preload is None
     assert preloads.curvature_matrix_counts is None
 
-    # LEq's blurred mapping matrices are different thus no preloading.
+    # Inversion's blurred mapping matrices are different thus no preloading.
 
     operated_mapping_matrix_0 = np.array(
         [[0.0, 1.0, 0.0], [1.0, 0.0, 0.0], [0.0, 0.0, 1.0]]
@@ -219,11 +219,11 @@ def test__set_operated_mapping_matrix_with_preloads():
         [[0.0, 0.0, 1.0], [1.0, 0.0, 0.0], [0.0, 0.0, 1.0]]
     )
 
-    leq_0 = aa.m.MockLEqImaging(operated_mapping_matrix=operated_mapping_matrix_0)
-    leq_1 = aa.m.MockLEqImaging(operated_mapping_matrix=operated_mapping_matrix_1)
+    inversion_0 = aa.m.MockInversionImaging(operated_mapping_matrix=operated_mapping_matrix_0)
+    inversion_1 = aa.m.MockInversionImaging(operated_mapping_matrix=operated_mapping_matrix_1)
 
-    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_0))
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_0 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_0))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(
         operated_mapping_matrix=1,
@@ -236,19 +236,19 @@ def test__set_operated_mapping_matrix_with_preloads():
     assert preloads.curvature_matrix_preload is None
     assert preloads.curvature_matrix_counts is None
 
-    # LEq's blurred mapping matrices are the same therefore preload it and the curvature sparse terms.
+    # Inversion's blurred mapping matrices are the same therefore preload it and the curvature sparse terms.
 
-    leq_0 = aa.m.MockLEqImaging(operated_mapping_matrix=operated_mapping_matrix_0)
-    leq_1 = aa.m.MockLEqImaging(operated_mapping_matrix=operated_mapping_matrix_0)
+    inversion_0 = aa.m.MockInversionImaging(operated_mapping_matrix=operated_mapping_matrix_0)
+    inversion_1 = aa.m.MockInversionImaging(operated_mapping_matrix=operated_mapping_matrix_0)
 
     inversion_0 = aa.m.MockInversion(
-        leq=leq_0,
+        inversion=inversion_0,
         curvature_matrix_preload=curvature_matrix_preload,
         curvature_matrix_counts=curvature_matrix_counts,
     )
 
     fit_0 = aa.m.MockFitImaging(inversion=inversion_0)
-    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(leq=leq_1))
+    fit_1 = aa.m.MockFitImaging(inversion=aa.m.MockInversion(inversion=inversion_1))
 
     preloads = aa.Preloads(
         operated_mapping_matrix=1,
@@ -296,7 +296,7 @@ def test__set_curvature_matrix():
 
     assert preloads.curvature_matrix is None
 
-    # LEq's curvature matrices are the same therefore preload it and the curvature sparse terms.
+    # Inversion's curvature matrices are the same therefore preload it and the curvature sparse terms.
 
     preloads = aa.Preloads(curvature_matrix=2)
 
@@ -329,7 +329,7 @@ def test__set_regularization_matrix_and_term():
 
     # Inversion's log_det_regularization_matrix_term are different thus no preloading.
 
-    leq = aa.m.MockLEq(linear_obj_list=[None])
+    inversion = aa.m.MockInversion(linear_obj_list=[None])
 
     linear_obj_reg = aa.m.MockLinearObjReg(regularization=regularization)
 
@@ -337,14 +337,14 @@ def test__set_regularization_matrix_and_term():
         inversion=aa.m.MockInversion(
             log_det_regularization_matrix_term=0,
             linear_obj_reg_list=[linear_obj_reg],
-            leq=leq,
+            inversion=inversion,
         )
     )
     fit_1 = aa.m.MockFitImaging(
         inversion=aa.m.MockInversion(
             log_det_regularization_matrix_term=1,
             linear_obj_reg_list=[linear_obj_reg],
-            leq=leq,
+            inversion=inversion,
         )
     )
 
@@ -354,13 +354,13 @@ def test__set_regularization_matrix_and_term():
     assert preloads.regularization_matrix is None
     assert preloads.log_det_regularization_matrix_term is None
 
-    # LEq's blurred mapping matrices are the same therefore preload it and the curvature sparse terms.
+    # Inversion's blurred mapping matrices are the same therefore preload it and the curvature sparse terms.
 
     preloads = aa.Preloads(log_det_regularization_matrix_term=2)
 
     fit_0 = aa.m.MockFitImaging(
         inversion=aa.m.MockInversion(
-            leq=aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()]),
+            inversion=aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()]),
             log_det_regularization_matrix_term=1,
             linear_obj_reg_list=[linear_obj_reg],
             preloads=preloads,
@@ -368,7 +368,7 @@ def test__set_regularization_matrix_and_term():
     )
     fit_1 = aa.m.MockFitImaging(
         inversion=aa.m.MockInversion(
-            leq=aa.m.MockLEq(linear_obj_list=[aa.m.MockMapper()]),
+            inversion=aa.m.MockInversion(linear_obj_list=[aa.m.MockMapper()]),
             log_det_regularization_matrix_term=1,
             linear_obj_reg_list=[linear_obj_reg],
             preloads=preloads,
